refactor: add NonProfitsSection to CausesPage
