+++ src/pages/donations/CausesPage/index.tsx
@@ -14,7 +14,6 @@ import { MODAL_TYPES } from "contexts/modalContext/helpers";
 import { getLocalStorageItem, setLocalStorageItem } from "lib/localStorage";
 import { DONATION_TOAST_SEEN_AT_KEY } from "lib/localStorage/constants";
 import { today } from "lib/dateTodayFormatter";
-import Spinner from "components/atomics/Spinner";
 import useVoucher from "hooks/useVoucher";
 import { useCausesContext } from "contexts/causesContext";
 import { logEvent, track } from "@amplitude/analytics-browser";
@@ -29,9 +28,9 @@ import UserSupportBanner from "components/moleculars/banners/UserSupportBanner";
 import useAvoidBackButton from "hooks/useAvoidBackButton";
 import { useCauseDonationContext } from "contexts/causeDonationContext";
 import { useNonProfitsContext } from "contexts/nonProfitsContext";
+import NonProfitsSection from "pages/donations/CausesPage/NonProfitsSection";
 import * as S from "./styles";
 import ContributionNotification from "./ContributionNotification";
-import NonProfitsList from "./NonProfitsList";
 import { LocationStateType } from "./LocationStateType";
 import ChooseCauseModal from "./ChooseCauseModal";
 import ContributionSection from "./ContributionSection";
@@ -43,9 +42,8 @@ function CausesPage(): JSX.Element {
 
   const { causesWithPoolBalance, isLoading: isLoadingCauses } =
     useCausesContext();
-  const { nonProfitsWithPoolBalance, isLoading: isLoadingNonProfits } =
-    useNonProfitsContext();
-  const { chosenCause, chooseCauseModalVisible } = useCauseDonationContext();
+  const { nonProfitsWithPoolBalance } = useNonProfitsContext();
+  const { chooseCauseModalVisible } = useCauseDonationContext();
 
   const { t } = useTranslation("translation", {
     keyPrefix: "donations.causesPage",
@@ -153,36 +151,6 @@ function CausesPage(): JSX.Element {
     }
   }, [chooseCauseModalVisible]);
 
-  const nonProfitsFilter = () => {
-    if (chosenCause) {
-      return (
-        nonProfitsWithPoolBalance?.filter(
-          (nonProfit) => nonProfit.cause?.id === chosenCause.id,
-        ) || []
-      );
-    }
-    return nonProfitsWithPoolBalance || [];
-  };
-
-  const sortNonProfits = () => {
-    const filteredNonProfits = nonProfitsFilter();
-    const sorted = filteredNonProfits?.sort((a, b) => {
-      const causeAIndex = causesWithPoolBalance.findIndex(
-        (cause) => cause.id === a.cause.id,
-      );
-      const causeBIndex = causesWithPoolBalance.findIndex(
-        (cause) => cause.id === b.cause.id,
-      );
-
-      return causeAIndex - causeBIndex;
-    });
-    return sorted;
-  };
-
-  useEffect(() => {
-    sortNonProfits();
-  }, [chosenCause]);
-
   useAvoidBackButton();
 
   return (
@@ -210,18 +178,7 @@ function CausesPage(): JSX.Element {
         <ContributionNotification />
         <CausesSelectSection />
 
-        {isLoadingNonProfits ? (
-          <Spinner size="26" />
-        ) : (
-          nonProfitsWithPoolBalance && (
-            <S.NonProfitsContainer>
-              <NonProfitsList
-                nonProfits={sortNonProfits()}
-                canDonate={canDonate}
-              />
-            </S.NonProfitsContainer>
-          )
-        )}
+        <NonProfitsSection canDonate={canDonate} />
         {isMobile && (
           <S.TooltipSection>
             <Tooltip
